requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
qdrant-client
sentence-transformers>=2.6.0
huggingface-hub>=0.20.0
groq
pypdf
python-multipart
requests
# torch==2.7.1+cpu